Always run initialization code on browser start
This seems to work more consistently than the existing code in Firefox.
Particularly, I could generally reproduce the extension not working if I quit
and restart Firefox Dev multiple times.

With this change, the extension loads on every browser startup, which
seems like the desired behavior, and also should be reinitialized on
upgrades.