FF: Use alarms to keep service worker running
Follow-up to #295, which doesn't seem to be enough to guarantee a good
search experience the first time typing.

This alarm approach seems to be not recommended, but it works well and
can be limited to just run on Firefox where it's an issue.

Without a much larger refactor or extensive testing, I think this is
truly the best short term solution to using the extension on Firefox!